location/grounding_dino/requirements.txt
# GroundingDINO
torch
pillow
transformers